requirements.txt
playwright==1.49.1
python-dotenv==1.0.1
pyyaml==6.0.2
aiohttp==3.11.11
apscheduler==3.10.4
aiosqlite==0.20.0
Jinja2==3.1.4
Pillow==11.1.0
